@@ -1,11 +1,12 @@
+from creart import create
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from graia.amnesia.builtins.uvicorn import UvicornService
+from graia.amnesia.builtins.asgi import UvicornASGIService
 from graia.broadcast import Broadcast
 from graia.saya import Saya
 from graia.saya.builtins.broadcast import BroadcastBehaviour
 from httpx import AsyncClient
-from launart import Launart, Launchable
+from launart import Launart, Service
 
 from graiax.fastapi import FastAPIBehaviour, FastAPIService
 from graiax.fastapi.saya.schema import RouteSchema
@@ -23,11 +24,11 @@ fastapi.add_middleware(
 
 @fastapi.get("/main")
 async def main():
-    return "main"
+    return "@fastapi.get('/main')"
 
 
 def test_main():
-    class MainLaunchable(Launchable):
+    class MainLaunchable(Service):
         id = "test"
 
         @property
@@ -42,45 +43,34 @@ def test_main():
             async with self.stage("blocking"):
                 with saya.module_context():
                     channel = saya.require("tests.schema")
-                assert (
-                    len(
-                        [
-                            c
-                            for c in channel.content
-                            if isinstance(c.metaclass, RouteSchema)
-                        ]
-                    )
-                    == 3
-                )
+
+                assert len([c for c in channel.content if isinstance(c.metaclass, RouteSchema)]) == 4
+
                 async with AsyncClient() as client:
-                    response = await client.get("http://127.0.0.1:8000/")
+                    response = await client.get("http://127.0.0.1:9000/")
                     assert response.json() == {"code": 200, "message": "Hello World!"}
-                    response = await client.get("http://127.0.0.1:8000/nothing")
-                    assert response.status_code == 404
-                with saya.module_context():
-                    saya.reload_channel(channel)
-                assert (
-                    len(
-                        [
-                            c
-                            for c in channel.content
-                            if isinstance(c.metaclass, RouteSchema)
-                        ]
-                    )
-                    == 3
-                )
-                async with AsyncClient() as client:
-                    response = await client.get("http://127.0.0.1:8000/")
-                    assert response.json() == {"code": 200, "message": "Hello World!"}
-                    response = await client.get("http://127.0.0.1:8000/nothing")
+                    response = await client.get("http://127.0.0.1:9000/nothing")
                     assert response.status_code == 404
 
-    launart = Launart()
-    bcc = Broadcast()
-    saya = Saya(bcc)
+                with saya.module_context():
+                    saya.reload_channel(channel)
+
+                assert len([c for c in channel.content if isinstance(c.metaclass, RouteSchema)]) == 4
+
+                async with AsyncClient() as client:
+                    response = await client.get("http://127.0.0.1:9000/")
+                    assert response.json() == {"code": 200, "message": "Hello World!"}
+                    response = await client.get("http://127.0.0.1:9000/nothing")
+                    assert response.status_code == 404
+
+                # launart.status.exiting = True
+
+    launart = create(Launart)
+    bcc = create(Broadcast)
+    saya = create(Saya)
     saya.install_behaviours(FastAPIBehaviour(fastapi))
     saya.install_behaviours(BroadcastBehaviour(bcc))
-    launart.add_service(FastAPIService(fastapi))
-    launart.add_service(UvicornService())
-    launart.add_launchable(MainLaunchable())
+    launart.add_component(FastAPIService(fastapi))
+    launart.add_component(UvicornASGIService("127.0.0.1", 9000, {"": fastapi}))  # type: ignore
+    launart.add_component(MainLaunchable())
     launart.launch_blocking()
